@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "description": "Report on status of Keyman projects",
   "main": "dist/server/code.js",
+  "type": "module",
   "scripts": {
     "build": "tsc",
     "start": "run-script-os",
@@ -12,8 +13,8 @@
   "author": "Marc Durdin",
   "license": "MIT",
   "dependencies": {
-    "@octokit/core": "^6.1.5",
-    "@octokit/plugin-rest-endpoint-methods": "^5.8.0",
+    "@octokit/core": "^5.2.0",
+    "@octokit/plugin-rest-endpoint-methods": "13.3.2-cjs.1",
     "@octokit/rest": "^21.1.1",
     "@sentry/node": "^9.8.0",
     "@sentry/tracing": "^7.120.3",
@@ -27,9 +28,9 @@
     "ws": "^7.4.6"
   },
   "devDependencies": {
-    "@octokit/types": "^6.25.0",
+    "@octokit/types": "^12.3.0",
     "@types/deep-equal": "^1.0.1",
-    "@types/express": "^4.17.11",
+    "@types/express": "^5.0.1",
     "@types/jest": "^27.0.1",
     "@types/node": "^16.7.1",
     "@types/parse-link-header": "^1.0.0",
@@ -39,6 +40,6 @@
     "mocha": "^9.2.2",
     "run-script-os": "^1.1.6",
     "tsc-watch": "^4.2.9",
-    "typescript": "^4.2.3"
+    "typescript": "^5.8.3"
   }
 }
