@@ -8,11 +8,13 @@
   This will emit markdown for the current test artifacts for that PR
 */
 
+import { exit } from "node:process";
+
 import { ProbotOctokit } from "probot";
-import { exit } from "process";
-import ManualTestParser from "../../shared/manual-test/manual-test-parser";
-import { ManualTestProtocol } from "../../shared/manual-test/manual-test-protocols";
-import { getArtifactLinksComment } from "./artifact-links-comment";
+
+import ManualTestParser from "../../shared/manual-test/manual-test-parser.js";
+import { ManualTestProtocol } from "../../shared/manual-test/manual-test-protocols.js";
+import { getArtifactLinksComment } from "./artifact-links-comment.js";
 
 const pr = 6849;
 const is_pull_request = true;
