@@ -1,7 +1,7 @@
 
-import httppost from '../../util/httppost';
-import { github_token } from '../../identity/github';
-import { logGitHubRateLimit } from '../../util/github-rate-limit';
+import httppost from '../../util/httppost.js';
+import { github_token } from '../../identity/github.js';
+import { logGitHubRateLimit } from '../../util/github-rate-limit.js';
 
 export default {
 
