@@ -1,8 +1,8 @@
 
-import httppost from '../../util/httppost';
-import { github_token } from '../../identity/github';
-import { logGitHubRateLimit } from '../../util/github-rate-limit';
-// import { getCurrentSprint } from '../../current-sprint';
+import httppost from '../../util/httppost.js';
+import { github_token } from '../../identity/github.js';
+import { logGitHubRateLimit } from '../../util/github-rate-limit.js';
+// import { getCurrentSprint } from '../../current-sprint.js';
 
 export default {
 
